requirements.txt
Flask
Flask-Cors
opencv-python
numpy
face-recognition
dlib
Pillow
scikit-learn
pandas
SQLAlchemy
pytest
gunicorn
PyJWT
# Optional
mysql-connector-python
python-dotenv
